Move OpenCode into extensions layout
