+++ extensions/opencode/src/client.ts
