+++ extensions/opencode/src/index.ts
