+++ extensions/opencode/src/proxy.ts
