+++ extensions/opencode/src/types.ts
