+++ extensions/opencode/tests/eviction.test.ts
